backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
tensorflow==2.15.0
python-multipart
jinja2
email-validator
passlib
bcrypt
PyJWT
reportlab
python-dotenv
matplotlib
seaborn
google-generativeai

pydantic>=2.5
typing-extensions